Split build graph target lowering
